Fixed MSVC warnings and a bug in GetUdpMasterserverForServer (a semicolon after if)
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@4164 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ include/CGameScript.h
@@ -212,7 +212,7 @@ public:
 	
 	const gs_worm_t	*getWorm()	const	{ return &Worm; }
 
-	int		getProjectileCount() const	{ return projectiles.size(); }
+	size_t	getProjectileCount() const	{ return projectiles.size(); }
 	
 	bool		Compile(const std::string& dir);
